Enabled warning as errors to all projects except app Fixed some java warnings Ignored unused result warnings in cmake

+++ test-app/app/src/main/java/com/tns/tests/Button1.java
@@ -160,7 +160,7 @@ public class Button1 extends ButtonBase {
     }
 
     public String callMethod2WithInt() {
-        return method2((int)1);
+        return method2(1);
     }
 
     public String callMethod2WithByte() {
